backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
Pillow==10.4.0
opencv-python==4.10.0.84
numpy==2.1.3
python-multipart==0.0.9
requests==2.32.3
